Fix empty `vendor/` directory
The `vendor/` directory is empty during Jenkins runes.

Since we are running Docker-in-Docker, the volume mount used during the
`vendor-install-deps` function is actually mounted from the Jenkins
host instead of the first Docker container. This means the `skopeo.tar`
is saved outside of the build directory.

This fixes the issue by ensuring we mount a path on the host into both
containers, and pass that path to `vendor-install-deps`.

Lastly:
- `vendor-install-deps` now uses less arguments, so it is
more simple to use
- `vendor-install-deps` now returns 1 if the tar is missing