Bootstrap `NEXUS_REGISTRY`
Our docs set `localhost` for the `NEXUS_URL`, because during the time
the `upload.sh` script is invoked DNS is not ready.

With DNS not being ready, `registry` won't work either. The `upload.sh`
script fails when it gets to Docker images.

Instead of having our users set two env. vars, just infer that if
`NEXUS_URL` is localhost that we should set `NEXUS_REGISTRY` to
localhost.